changed pycord.py to main.py
